Create basic graph generator
This reads in each line of the input file and generates the
corresponding tree graph. It takes the level of indentation to key both
the parent-child relationship and the graph rank. Multiple trees can be
rendered by a single file, each tree starting with a line containing
"tree". Labels are drawn on the nodes and multiline labels use " - " as
the line separator.